+++ src/Cake.AWS.Route53/Extensions/Context/CakeContextExtensions.cs
@@ -1,9 +1,9 @@
 ﻿#region Using Statements
-    using System;
+using System;
 
-    using Cake.Core;
+using Cake.Core;
 
-    using Amazon;
+using Amazon;
 #endregion